Refactor to response correct status error when topic ID does not exist

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -22,6 +22,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -62,25 +63,40 @@ public class TopicosController {
     }
     
     @GetMapping("/{id}")
-    public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-    	Topico topico = topicoRepository.getOne(id);
-    	return new DetalhesDoTopicoDto(topico); 
+    public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+    	Optional<Topico> topico = topicoRepository.findById(id);
+    	
+    	if(topico.isPresent()) {
+    		return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));     		
+    	}
+    	return ResponseEntity.notFound().build();
     }
     
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicosDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizarTopicoForm form) {
-    	form.atualizar(id, topicoRepository);
+    	Optional<Topico> optionalTopic = topicoRepository.findById(id);
     	
-    	Topico topico = form.atualizar(id, topicoRepository);
-    	return ResponseEntity.ok(new TopicosDto(topico));
+    	if(optionalTopic.isPresent()) {
+    		form.atualizar(id, topicoRepository);
+    		
+    		Topico topico = form.atualizar(id, topicoRepository);
+    		return ResponseEntity.ok(new TopicosDto(topico));    		
+    	}
+    	
+    	return ResponseEntity.notFound().build();
     }
     
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> remover(@PathVariable Long id){
-    	topicoRepository.deleteById(id);
+    	Optional<Topico> topico = topicoRepository.findById(id);
     	
-    	return ResponseEntity.ok().build();
+    	if(topico.isPresent()) {
+    		topicoRepository.deleteById(id);
+    		
+    		return ResponseEntity.ok().build();    		
+    	}
+    	return ResponseEntity.notFound().build();
     }
 }
